requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --config=pyproject.toml requirements/requirements.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via litellm
aiosignal==1.3.1
    # via aiohttp
amqp==5.2.0
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyascii==0.3.2
    # via wagtail
anyio==4.4.0
    # via
    #   httpx
    #   openai
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
attrs==24.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via wagtail
billiard==4.2.0
    # via celery
brotli==1.1.0
    # via whitenoise
celery[redis]==5.4.0
    # via
    #   -r requirements/requirements.in
    #   django-celery-beat
celery-progress==0.4
    # via -r requirements/requirements.in
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.17.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   litellm
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.3
    # via django-celery-beat
cryptography==43.0.0
    # via
    #   fido2
    #   pyjwt
defusedxml==0.7.1
    # via willow
distro==1.9.0
    # via openai
dj-rest-auth==6.0.0
    # via -r requirements/requirements.in
dj-stripe==2.8.4
    # via -r requirements/requirements.in
django==5.1.1
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-health-check
    #   django-hijack
    #   django-modelcluster
    #   django-otp
    #   django-permissionedforms
    #   django-taggit
    #   django-timezone-field
    #   django-treebeard
    #   django-waffle
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   laces
    #   wagtail
django-allauth[mfa,socialaccount]==64.1.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==11.1
    # via -r requirements/requirements.in
django-celery-beat==2.7.0
    # via -r requirements/requirements.in
django-cors-headers==4.4.0
    # via -r requirements/requirements.in
django-environ==0.11.2
    # via -r requirements/requirements.in
django-filter==24.3
    # via wagtail
django-health-check==3.18.3
    # via -r requirements/requirements.in
django-hijack==3.6.0
    # via -r requirements/requirements.in
django-modelcluster==6.3
    # via wagtail
django-otp==1.5.2
    # via -r requirements/requirements.in
django-permissionedforms==0.1
    # via wagtail
django-taggit==5.0.1
    # via wagtail
django-timezone-field==7.0
    # via django-celery-beat
django-treebeard==4.7.1
    # via wagtail
django-waffle==4.1.0
    # via -r requirements/requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements/requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   wagtail
djangorestframework-api-key==3.0.0
    # via -r requirements/requirements.in
djangorestframework-simplejwt==5.3.1
    # via -r requirements/requirements.in
docutils==0.21.2
    # via -r requirements/requirements.in
draftjs-exporter==5.0.0
    # via wagtail
drf-spectacular==0.27.2
    # via -r requirements/requirements.in
et-xmlfile==1.1.0
    # via openpyxl
fido2==1.1.3
    # via django-allauth
filelock==3.15.4
    # via huggingface-hub
filetype==1.2.0
    # via willow
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.1
    # via huggingface-hub
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
huggingface-hub==0.24.6
    # via tokenizers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.4.0
    # via litellm
inflection==0.5.1
    # via drf-spectacular
jinja2==3.1.4
    # via litellm
jiter==0.5.0
    # via openai
jsonschema==4.23.0
    # via
    #   drf-spectacular
    #   litellm
jsonschema-specifications==2023.12.1
    # via jsonschema
kombu==5.4.0
    # via celery
l18n==2021.3
    # via wagtail
laces==0.1.1
    # via wagtail
litellm==1.44.4
    # via -r requirements/requirements.in
mailchimp3==3.0.21
    # via -r requirements/requirements.in
markdown==3.7
    # via -r requirements/requirements.in
markupsafe==2.1.5
    # via jinja2
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.42.0
    # via
    #   -r requirements/requirements.in
    #   litellm
openpyxl==3.1.5
    # via wagtail
packaging==24.1
    # via
    #   djangorestframework-api-key
    #   huggingface-hub
pillow==10.4.0
    # via
    #   pillow-heif
    #   wagtail
pillow-heif==0.18.0
    # via willow
prompt-toolkit==3.0.47
    # via click-repl
psycopg2-binary==2.9.9
    # via -r requirements/requirements.in
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   litellm
    #   openai
pydantic-core==2.20.1
    # via pydantic
pyjwt[crypto]==2.9.0
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
pypng==0.20220715.0
    # via qrcode
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   python-crontab
python-dotenv==1.0.1
    # via litellm
pytz==2024.1
    # via
    #   django-modelcluster
    #   l18n
pyyaml==6.0.2
    # via
    #   drf-spectacular
    #   huggingface-hub
qrcode==7.4.2
    # via django-allauth
redis==5.0.8
    # via celery
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   django-allauth
    #   django-anymail
    #   huggingface-hub
    #   litellm
    #   mailchimp3
    #   requests-oauthlib
    #   stripe
    #   tiktoken
    #   wagtail
requests-oauthlib==2.0.0
    # via django-allauth
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
sentry-sdk==2.13.0
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   l18n
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.6
    # via beautifulsoup4
sqlparse==0.5.1
    # via django
stripe==4.2.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
telepath==0.3.1
    # via wagtail
tiktoken==0.7.0
    # via litellm
tokenizers==0.20.0
    # via litellm
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   qrcode
tzdata==2024.1
    # via
    #   celery
    #   django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.2
    # via
    #   django-anymail
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wagtail==6.2.1
    # via -r requirements/requirements.in
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise[brotli]==6.7.0
    # via -r requirements/requirements.in
willow[heif]==1.8.0
    # via wagtail
yarl==1.9.4
    # via aiohttp
zipp==3.20.0
    # via importlib-metadata
